refactor: extract IPC dispatch and CWD sync to commands/ipc_dispatch.rs (#127 step 8b)
Move handle_lx_message*, CWD sync helpers (resolve_target_terminals,
filter_targets_*, write_cd_to_group_terminals, build_sync_cd_command,
mark_propagated, etc.) and ~100 related tests to ipc_dispatch.rs.
